% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ABV}
\alias{ABV}
\title{Calculates ABV-index}
\usage{
ABV(data, bin_width, ID, trial, grouping, value, label, type)
}
\arguments{
\item{data}{A dataset in tabular format.}

\item{bin_width}{Single integer value which gives the number of consecutive trials handled in one bin.}

\item{ID}{A  character vector identifying the name of column for subjects.}

\item{trial}{A single character vector determining the name of your measured variable, like reaction time.}

\item{grouping}{A list of strings determining by which cathegorical variable you would like to group your data.}

\item{value}{A single character vector identifying the column containing your measured variable like reaction time as RT.}

\item{label}{A single character vector identifying the column containing trial types.}

\item{type}{A  character vector identifying the names of your trial types. First column shall address incongruent trials.}

\item{shaker}{Logical, by default is FALSE. If TRUE the function reshapes the data as in a separate shaker() function.}
}
\description{
\code{ABV} calculates Attention Bias Variability (ABV) index.
}
\examples{
ABV(example2, grouping = c("year", "subject"), trial = "trial", bin_width = 3, value = "RT", label = "label", type = c("Incongruent", "Congruent"), ID = "subject")
}
\seealso{
\code{\link{winzer}} for winsorizing your data, \code{\link{shaker}} for reshape your data.
}
