% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example}
\alias{example}
\title{Dataset 1. - "example"}
\format{A dataset with 12 variables and 20 observations
\describe{
  \item{year}{The simulated year of the data collection}
  \item{subject}{Subject ID}
  \item{trial}{10 trial variables for simulated reaction time in a dot-probe test; measured in ms}
}}
\usage{
example
}
\description{
A simulated reaction time dataset for demonstrating the shaker() function in the abvyr package
}
\keyword{datasets}
