% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example2}
\alias{example2}
\title{Dataset 2. - "example2"}
\format{A dataset with 5 variables and 200 observations
\describe{
  \item{year}{The simulated year of the data collection}
  \item{subject}{Subject ID}
  \item{trial}{trial name}
  \item{RT}{reaction time to a dot-prove test; measured in ms}
  \item{label}{flag variable for indicating if a trial is congruent or incongruent to the emotional stimuli}
}}
\usage{
example2
}
\description{
A simulated reaction time dataset for demonstrating the ABV-index calculation in the abvyr package
}
\keyword{datasets}
