% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro.R
\docType{package}
\name{RRT}
\alias{RRT}
\alias{RRT-package}
\title{REAC: Robust Reaction Times Analysis toolset
A package for attention research related reaction times analysis}
\description{
The REAC package provides tools for preparing your data for additional calculations related to
ABV-index (Attention Bias Variability) and data cleaning techniques such as winsorizing your data.
}
\section{abvyr package functions}{

\code{\link{shaker}} to shape your data

\code{\link{winzer}} to winsorize your data

\code{\link{ABV}} to calculate ABV index
}

