% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{shaker}
\alias{shaker}
\title{Tidies up the dataset and prepares later data analysis}
\usage{
shaker(data, grouping, gather, value, ...)
}
\arguments{
\item{data}{A dataset in tabular format.}

\item{grouping}{A list of strings determining by which cathegorical variable you would like to group your data.}

\item{gather}{A single character vector determining by which cathegorical variable you would like to gather your data.}

\item{value}{A single character vector identifying the column containing your measured variable like reaction time as RT.}
}
\description{
\code{shaker} prepares data for later analysis like calculating Attention Bias Variability (ABV) index.
}
\examples{
shaker(example, grouping = c("year", "subject"), gather = "trial", value = "RT")
}
\seealso{
\code{\link{winzer}} for winsorizing your data, \code{\link{ABV}} for calculating ABV-index
}
