% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{winzer}
\alias{winzer}
\title{Winsorizes your data}
\usage{
winzer(data, x, y, z, grouping, value, label, ...)
}
\arguments{
\item{data}{A dataset in tabular format.}

\item{x}{Single value declaration for lower quantile of winsorizing.}

\item{y}{Single value declaration for upper quantile of winsorizing.}

\item{z}{Single value declaration for IQR multiplier.}

\item{grouping}{A list of strings determining by which cathegorical variable you would like to group your data.}

\item{value}{A single character vector identifying the column containing your measured variable like reaction time as RT.}

\item{label}{A string name indicating the columns containing trials or observation points.}
}
\description{
\code{winzer} performs winsorization on your data.
}
\examples{
winzer(example, grouping = c("year", "subject"), x = 0.25, y = 0.75, z = 1.5, label = "trial", value = "RT")
}
\seealso{
\code{\link{ABV}} for ABV-index, \code{\link{shaker}} for reshape your data.
For extended winsorization see \code{\link[WRS2]{winvar}} and the win function family of the WRS package.
}
